Don't use du -x to calculate image size
du -x means only check the current file system. There is currently a bug
in overlayfs where files report to be on different filesystems and so
are not being counted by dib when building images within a docker
container.

It appears that the -x flag has been there since the beginning of this
file [1] and carried forward, however at this point in dib there should
not be anything mounted in the folder so hopefully not make a
difference.

There is a patch that should solve this issue [2] but who knows how long
it will take to land on builders.

[1] https://review.openstack.org/#/c/25866/
[2] https://patchwork.kernel.org/patch/9706629/

Closes-Bug: #1702382
Change-Id: I563dbb84cfefaf7eabd2f1af5ba04be9648f7155